requirements.txt
Flask==2.3.3
Pillow==10.0.1
requests==2.31.0
numpy==1.24.3
python-dotenv==1.0.0
scikit-learn==1.3.0
colorthief==0.2.1